requirements.txt
aiohttp==3.6.3
aiosqlite==0.15.0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
dateparser==0.7.6
discord.py==1.5.1
idna==2.10
multidict==4.7.6
mutagen==1.45.1
numpy==1.19.2
pandas==1.1.3
pyaml==20.4.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
PyNaCl==1.3.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
PyYAML==5.3.1
rapidfuzz==0.12.3
regex==2020.10.15
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
yarl==1.5.1
